Dia 12/08/2025 adicionando SWAGGER, exceções e coligando clientes a carros, iniciando também coligação entre gerente e estacionamneto (WOP)

+++ src/main/java/com/senai/CadastroEstacionamento/aplication/service/CarroService.java
@@ -2,6 +2,7 @@ package com.senai.CadastroEstacionamento.aplication.service;
 
 import com.senai.CadastroEstacionamento.aplication.dtos.CarroDto;
 import com.senai.CadastroEstacionamento.domain.entity.Carro;
+import com.senai.CadastroEstacionamento.domain.exception.IdDesconhecidoException;
 import com.senai.CadastroEstacionamento.domain.repository.CarroRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,12 +29,14 @@ public class CarroService {
                 .collect(Collectors.toList());
     }
 
-    public Optional<CarroDto> buscarPorId(Long id) {
-        return carroRepository.findById(id)
-                .map(CarroDto::toDto);
+    public Carro buscarPorId(Long id) {
+        return carroRepository.findById(id).orElseThrow(() -> new IdDesconhecidoException("Usuário não encontrado com o ID: " + id));
+
     }
 
     public boolean atualizar(Long id, CarroDto dto) {
+        carroRepository.findById(id).orElseThrow(() -> new IdDesconhecidoException("Usuário não encontrado com o ID: " + id));
+
         return carroRepository.findById(id).map(carro -> {
             Carro carroAtualizado = dto.fromDto();
 
@@ -43,10 +46,11 @@ public class CarroService {
     }
 
     public boolean deletar(Long id) {
-        return carroRepository.findById(id).map(carro -> {
-            carroRepository.deleteById(id);
-            return true;
-        }).orElse(false);
+        if (!carroRepository.existsById(id)){
+            throw new IdDesconhecidoException("Usuário não encontrado com o ID: " + id);
+        }
+        carroRepository.deleteById(id);
+        return true;
     }
 
 }
